refactoring of unittests
